# lib/api-zod/src/generated/types/consumptionEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Interphase Engineers - Project Control & Margin Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsumptionEntry {
  id: number;
  projectId: number;
  weekNumber: number;
  weekStart: Date;
  weekEnd: Date;
  itemName: string;
  estimatedQty: number;
  actualQty: number;
  /** Quantity returned after project completion */
  returnedQty: number;
  /** actualQty minus returnedQty — the quantity consumed net of returns */
  netQty: number;
  variance: number;
  variancePercent: number;
  unitCost: number;
  actualCost: number;
  /** netQty × unitCost — the accurate cost after accounting for returned material */
  netCost: number;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}