# lib/api-zod/src/generated/types/estimateRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Interphase Engineers - Project Control & Margin Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface EstimateRow {
  id: number;
  projectId: number;
  itemCategory: string;
  itemName: string;
  specification?: string;
  unit: string;
  estimatedQuantity: number;
  rate: number;
  cost: number;
  remarks?: string;
  createdAt: Date;
  updatedAt: Date;
}